# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Story {
  id: string;
  title: string;
  category: string;
  duration: number;
  ageMin: number;
  ageMax: number;
  description: string;
  thumbnailUrl?: string | null;
  audioUrl?: string | null;
  videoUrl?: string | null;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStory {
  id: string;
  title: string;
  category: string;
  duration: number;
  ageMin: number;
  ageMax: number;
  description: string;
  thumbnailUrl?: string;
  audioUrl?: string;
  videoUrl?: string;
  published?: boolean;
}

export interface UpdateStory {
  title?: string;
  category?: string;
  duration?: number;
  ageMin?: number;
  ageMax?: number;
  description?: string;
  thumbnailUrl?: string;
  audioUrl?: string;
  videoUrl?: string;
  published?: boolean;
}

export interface ErrorResponse {
  error: string;
}