# lib/api-zod/src/generated/types/updateStory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStory {
  title?: string;
  category?: string;
  duration?: number;
  ageMin?: number;
  ageMax?: number;
  description?: string;
  thumbnailUrl?: string;
  audioUrl?: string;
  videoUrl?: string;
  published?: boolean;
}